+++ src/lib/bson/document.ts
@@ -1,50 +1,66 @@
-import { diffSets, DiffSetsResult } from "../_helpers/diff-sets";
-import { createExtraKeysError } from "../errors/extra-keys";
-import { createMissingKeysError } from "../errors/missing-keys";
+import { createInvalidDocumentError } from "../errors/invalid-document";
 import { Serializer } from "../types";
-import { DocumentType, name as typeName } from "../types/document";
+import { DocumentType, name as typeName, PropertyDescriptor, renameKeys } from "../types/document";
 
 export function register(serializer: Serializer): void {
-  function write<T extends {}>(type: DocumentType<T>, val: T): {[key: string]: any} {
-    const keysDiff: DiffSetsResult<string> = diffSets(type.keys.keys(), Object.keys(val));
-    const result: {[key: string]: any} = {};
-    for (const key of keysDiff.commonKeys) {
-      if ((<any> val)[key] === undefined && type.properties[key].optional) {
-        continue;
+  function write<T extends {}>(type: DocumentType<T>, val: T): any {
+    const result: any = Object.create(null);
+    for (const [key, outKey] of renameKeys(type.properties, type.rename)) {
+      const descriptor: PropertyDescriptor<T[keyof T]> = type.properties[key];
+      const value: T[keyof T] = val[key];
+      if (value === undefined) {
+        Reflect.set(result, outKey, undefined);
+      } else {
+        Reflect.set(result, outKey, serializer.write(descriptor.type, value));
       }
-      const outKey: string = type.keys.get(key)!;
-      result[outKey] = serializer.write(type.properties[key].type, (<any> val)[key]);
-    }
-    return result;
-  }
-
-  function read<T>(type: DocumentType<T>, input: {[key: string]: any}): T {
-    const outKeysDiff: DiffSetsResult<string> = diffSets(type.outKeys.keys(), Object.keys(input));
-    const missingRequiredKeys: string[] = [...outKeysDiff.missingKeys].filter((outKey: string): boolean => {
-      return !type.properties[type.outKeys.get(outKey)!].optional;
-    });
-    if (missingRequiredKeys.length > 0) {
-      throw createMissingKeysError(missingRequiredKeys);
-    } else if (outKeysDiff.extraKeys.size > 0 && !type.ignoreExtraKeys) {
-      throw createExtraKeysError([...outKeysDiff.extraKeys]);
-    }
-
-    // TODO(demurgos): use Partial<T> once typedoc supports it
-    const result: any = {};
-    for (const outKey of outKeysDiff.commonKeys) {
-      const key: string = type.outKeys.get(outKey)!;
-      result[key] = serializer.read(type.properties[key].type, input[outKey]);
     }
     return result as T;
   }
 
-  function readTrusted<T>(type: DocumentType<T>, input: {[key: string]: any}): T {
-    const outKeysDiff: DiffSetsResult<string> = diffSets(type.outKeys.keys(), Object.keys(input));
-    // TODO(demurgos): use Partial<T> once typedoc supports it
-    const result: any = {};
-    for (const outKey of outKeysDiff.commonKeys) {
-      const key: string = type.outKeys.get(outKey)!;
-      result[key] = serializer.readTrusted(type.properties[key].type, input[outKey]);
+  function read<T extends {}>(type: DocumentType<T>, input: any): T {
+    const extra: Set<string> | undefined = type.ignoreExtraKeys ? undefined : new Set(Object.keys(input));
+    const missing: Set<string> = new Set();
+    const invalid: Map<keyof T, Error> = new Map();
+
+    const result: Partial<T> = Object.create(null);
+
+    for (const [key, outKey] of renameKeys(type.properties, type.rename)) {
+      if (extra !== undefined) {
+        extra.delete(outKey);
+      }
+      const descriptor: PropertyDescriptor<any> = type.properties[key];
+      const outValue: any = Reflect.get(input, outKey);
+      if (outValue === undefined) {
+        if (descriptor.optional) {
+          result[key] = undefined;
+        } else {
+          missing.add(key);
+        }
+        continue;
+      }
+      try {
+        result[key] = serializer.read(descriptor.type, outValue);
+      } catch (err) {
+        invalid.set(key, err);
+      }
+    }
+
+    if (extra !== undefined && extra.size > 0 || missing.size > 0 || invalid.size > 0) {
+      throw createInvalidDocumentError({extra, missing, invalid});
+    }
+    return result as T;
+  }
+
+  function readTrusted<T extends {}>(type: DocumentType<T>, input: any): T {
+    const result: Partial<T> = Object.create(null);
+    for (const [key, outKey] of renameKeys(type.properties, type.rename)) {
+      const descriptor: PropertyDescriptor<any> = type.properties[key];
+      const outValue: any = Reflect.get(input, outKey);
+      if (outValue === undefined) {
+        result[key] = undefined;
+      } else {
+        result[key] = serializer.readTrusted(descriptor.type, outValue);
+      }
     }
     return result as T;
   }