+++ src/lib/types/document.ts
@@ -1,40 +1,27 @@
 import { Incident } from "incident";
-import { diffSets, DiffSetsResult } from "../_helpers/diff-sets";
 import { lazyProperties } from "../_helpers/lazy-properties";
 import { CaseStyle, rename } from "../case-style";
-import { createExtraKeysError } from "../errors/extra-keys";
-import { createInvalidPropertyError } from "../errors/invalid-property";
+import { createInvalidDocumentError } from "../errors/invalid-document";
 import { createInvalidTypeError } from "../errors/invalid-type";
-import { createMissingKeysError } from "../errors/missing-keys";
 import { createNotImplementedError } from "../errors/not-implemented";
-import { createNullPropertyError } from "../errors/null-property";
+import { JSON_SERIALIZER } from "../json";
 import { Lazy, Type as KryoType, VersionedType } from "../types";
 
 export type Name = "document";
 export const name: Name = "document";
 export namespace json {
-  export interface Input {
-    [key: string]: any;
-  }
-
-  export interface Output {
-    [key: string]: any;
-  }
-
   export interface Type {
     name: Name;
-    notNan: boolean;
-    notInfinity: boolean;
   }
 }
 
-export interface Diff {
-  set: {[key: string]: any}; // val
-  update: {[key: string]: any}; // diff
-  unset: {[key: string]: any}; // val
+export interface Diff<T> {
+  set: {[P in keyof T]?: any}; // val
+  update: {[P in keyof T]?: any}; // diff
+  unset: {[P in keyof T]?: any}; // val
 }
 
-export interface Options<TypeKind extends KryoType<any>> {
+export interface DocumentTypeOptions<T> {
   /**
    * Do not throw error when the object contains extraneous keys.
    */
@@ -43,7 +30,7 @@ export interface Options<TypeKind extends KryoType<any>> {
   /**
    * A dictionary between a property name and its description.
    */
-  properties: {[key: string]: PropertyDescriptor<TypeKind>};
+  properties: {[P in keyof T]: PropertyDescriptor<KryoType<T[P]>>};
 
   /**
    * The keys of the serialized documents are renamed following the
@@ -52,7 +39,7 @@ export interface Options<TypeKind extends KryoType<any>> {
   rename?: CaseStyle;
 }
 
-export interface PropertyDescriptor<TypeKind extends KryoType<any>> {
+export interface PropertyDescriptor<MetaType extends KryoType<any>> {
   /**
    * Allows this property to be missing (undefined values throw errors).
    */
@@ -61,43 +48,45 @@ export interface PropertyDescriptor<TypeKind extends KryoType<any>> {
   /**
    * The type of this property.
    */
-  type: TypeKind;
+  type: MetaType;
 }
 
-export class DocumentType<T extends {}> implements VersionedType<T, json.Input, json.Output, Diff> {
+export interface DocumentTypeConstructor {
+  new<T>(options: Lazy<DocumentTypeOptions<T>>): DocumentType<T>;
+}
+
+export interface DocumentType<T> extends VersionedType<T, any, any, Diff<T>>, DocumentTypeOptions<T> {
+
+}
+
+// tslint:disable-next-line:variable-name
+export const DocumentType: DocumentTypeConstructor = class<T extends {}> {
   readonly name: Name = name;
-  readonly ignoreExtraKeys!: boolean;
-  readonly properties!: {
-    [key: string]: PropertyDescriptor<VersionedType<any, any, any, any>>;
-  };
+  readonly ignoreExtraKeys: boolean;
+  readonly properties: {[P in keyof T]: PropertyDescriptor<KryoType<T[P]>>};
   readonly rename?: CaseStyle;
 
   /**
    * Map from the document keys to the serialized names
    */
-    // TODO: Restrict visibility (public -> private)
-  public readonly keys!: Map<string, string>;
+  private readonly keys: Map<keyof T, string>;
 
-  /**
-   * Map from the serialized names to the document keys
-   */
-    // TODO: Restrict visibility (public -> private)
-  public readonly outKeys!: Map<string, string>;
+  private _options: Lazy<DocumentTypeOptions<T>>;
 
-  private _options: Lazy<Options<VersionedType<any, any, any, any>>>;
+  constructor(options: Lazy<DocumentTypeOptions<T>>) {
+    // TODO: Remove once TS 2.7 is better supported by editors
+    this.ignoreExtraKeys = <any> undefined;
+    this.properties = <any> undefined;
+    this.keys = <any> undefined;
 
-  constructor(options: Lazy<Options<VersionedType<any, any, any, any>>>, lazy?: boolean) {
     this._options = options;
-    if (lazy === undefined) {
-      lazy = typeof options === "function";
-    }
-    if (!lazy) {
+    if (typeof options !== "function") {
       this._applyOptions();
     } else {
       lazyProperties(
         this,
         this._applyOptions,
-        ["ignoreExtraKeys", "properties", "rename", "keys", "outKeys"],
+        ["ignoreExtraKeys", "properties", "rename", "keys"],
       );
     }
   }
@@ -110,126 +99,159 @@ export class DocumentType<T extends {}> implements VersionedType<T, json.Input, 
     throw createNotImplementedError("DocumentType#toJSON");
   }
 
-  readTrustedJson(input: json.Output): T {
-    const outKeysDiff: DiffSetsResult<string> = diffSets(this.outKeys.keys(), Object.keys(input));
-    // TODO(demurgos): use Partial<T> once typedoc supports it
-    const result: any = {};
-    for (const outKey of outKeysDiff.commonKeys) {
-      const key: string = this.outKeys.get(outKey)!;
-      result[key] = this.properties[key].type.readJson(input[outKey]);
+  readTrustedJson(input: any): T {
+    const result: Partial<T> = Object.create(null);
+    for (const [key, outKey] of this.keys) {
+      const descriptor: PropertyDescriptor<any> = this.properties[key];
+      const jsonValue: any = Reflect.get(input, outKey);
+      if (jsonValue === undefined) {
+        result[key] = undefined;
+      } else {
+        result[key] = JSON_SERIALIZER.readTrusted(descriptor.type, jsonValue);
+      }
     }
     return result as T;
   }
 
   readJson(input: any): T {
-    const outKeysDiff: DiffSetsResult<string> = diffSets(this.outKeys.keys(), Object.keys(input));
-    const missingRequiredKeys: string[] = [...outKeysDiff.missingKeys].filter((outKey: string): boolean => {
-      return !this.properties[this.outKeys.get(outKey)!].optional;
-    });
-    if (missingRequiredKeys.length > 0) {
-      throw createMissingKeysError(missingRequiredKeys);
-    } else if (outKeysDiff.extraKeys.size > 0 && !this.ignoreExtraKeys) {
-      throw createExtraKeysError([...outKeysDiff.extraKeys]);
+    const extra: Set<string> | undefined = this.ignoreExtraKeys ? undefined : new Set(Object.keys(input));
+    const missing: Set<string> = new Set();
+    const invalid: Map<keyof T, Error> = new Map();
+
+    const result: Partial<T> = Object.create(null);
+
+    for (const [key, outKey] of this.keys) {
+      if (extra !== undefined) {
+        extra.delete(outKey);
+      }
+      const descriptor: PropertyDescriptor<any> = this.properties[key];
+      const jsonValue: any = Reflect.get(input, outKey);
+      if (jsonValue === undefined) {
+        if (descriptor.optional) {
+          result[key] = undefined;
+        } else {
+          missing.add(key);
+        }
+        continue;
+      }
+      try {
+        result[key] = JSON_SERIALIZER.read(descriptor.type, jsonValue);
+      } catch (err) {
+        invalid.set(key, err);
+      }
     }
 
-    // TODO(demurgos): use Partial<T> once typedoc supports it
-    const result: any = {};
-    for (const outKey of outKeysDiff.commonKeys) {
-      const key: string = this.outKeys.get(outKey)!;
-      // TODO(demurgos): Check if the format is supported instead of casting to `any`
-      result[key] = this.properties[key].type.readJson(input[outKey]);
+    if (extra !== undefined && extra.size > 0 || missing.size > 0 || invalid.size > 0) {
+      throw createInvalidDocumentError({extra, missing, invalid});
     }
     return result as T;
   }
 
-  writeJson(val: T): json.Output {
-    const keysDiff: DiffSetsResult<string> = diffSets(this.keys.keys(), Object.keys(val));
-    const result: {[key: string]: any} = {};
-    for (const key of keysDiff.commonKeys) {
-      if ((<any> val)[key] === undefined && this.properties[key].optional) {
-        continue;
+  writeJson(val: T): any {
+    const result: any = Object.create(null);
+    for (const [key, outKey] of this.keys) {
+      const descriptor: PropertyDescriptor<T[keyof T]> = this.properties[key];
+      const value: T[keyof T] = val[key];
+      if (value === undefined) {
+        Reflect.set(result, outKey, undefined);
+      } else {
+        Reflect.set(result, outKey, JSON_SERIALIZER.write(descriptor.type, value));
       }
-      const outKey: string = this.keys.get(key)!;
-      result[outKey] = this.properties[key].type.writeJson((<any> val)[key]);
     }
-    return result;
+    return result as T;
   }
 
   testError(val: T): Error | undefined {
     if (typeof val !== "object" || val === null) {
       return createInvalidTypeError("object", val);
     }
-    const keysDiff: DiffSetsResult<string> = diffSets(this.keys.keys(), Object.keys(val));
-    const missingRequiredKeys: string[] = [...keysDiff.missingKeys].filter((key: string): boolean => {
-      return !this.properties[key].optional;
-    });
-    if (missingRequiredKeys.length > 0) {
-      return createMissingKeysError(missingRequiredKeys);
-    } else if (keysDiff.extraKeys.size > 0 && !this.ignoreExtraKeys) {
-      return createExtraKeysError([...keysDiff.extraKeys]);
+
+    const extra: Set<string> | undefined = this.ignoreExtraKeys ? undefined : new Set(Object.keys(val));
+    const missing: Set<string> = new Set();
+    const invalid: Map<keyof T, Error> = new Map();
+
+    for (const key in this.properties) {
+      if (extra !== undefined) {
+        extra.delete(key);
+      }
+      const descriptor: PropertyDescriptor<any> = this.properties[key];
+      const propertyValue: any = val[key];
+      if (propertyValue === undefined) {
+        if (!descriptor.optional) {
+          missing.add(key);
+        }
+        continue;
+      }
+      const error: Error | undefined = descriptor.type.testError(propertyValue);
+      if (error !== undefined) {
+        invalid.set(key, error);
+      }
     }
 
-    for (const key of keysDiff.commonKeys) {
-      const member: any = (<any> val)[key];
-      const descriptor: PropertyDescriptor<KryoType<any>> = this.properties[key];
-      if (member === undefined) {
-        if (descriptor.optional) {
-          continue;
-        } else {
-          return createNullPropertyError(key);
-        }
-      }
-      const error: Error | undefined = descriptor.type.testError(member);
-      if (error !== undefined) {
-        return createInvalidPropertyError(key, member);
-      }
+    if (extra !== undefined && extra.size > 0 || missing.size > 0 || invalid.size > 0) {
+      return createInvalidDocumentError({extra, missing, invalid});
     }
     return undefined;
   }
 
   test(val: T): val is T {
-    return this.testError(val) === undefined;
+    if (typeof val !== "object" || val === null) {
+      return false;
+    }
+
+    const extra: Set<string> | undefined = this.ignoreExtraKeys ? undefined : new Set(Object.keys(val));
+
+    for (const key in this.properties) {
+      if (extra !== undefined) {
+        extra.delete(key);
+      }
+      const descriptor: PropertyDescriptor<any> = this.properties[key];
+      const propertyValue: any = val[key];
+      if (propertyValue === undefined) {
+        if (!descriptor.optional) {
+          return false;
+        }
+      } else if (!descriptor.type.test(propertyValue)) {
+        return false;
+      }
+    }
+
+    return extra === undefined || extra.size === 0;
   }
 
   equals(val1: T, val2: T): boolean {
     for (const key in this.properties) {
       const descriptor: PropertyDescriptor<KryoType<any>> = this.properties[key];
-      const member1: any = (<any> val1)[key];
-      const member2: any = (<any> val2)[key];
       if (!descriptor.optional) {
-        if (!descriptor.type.equals(member1, member2)) {
+        if (!descriptor.type.equals(val1[key], val2[key])) {
           return false;
         }
-      } else {
-        if (val1.hasOwnProperty(key) && val2.hasOwnProperty(key)) {
-          if (!descriptor.type.equals(member1, member2)) {
-            return false;
-          }
-        } else if (!(member1 === undefined && member2 === undefined)) {
-          return false;
-        }
+        continue;
+      }
+      if (val1[key] === undefined && val2[key] === undefined) {
+        continue;
+      }
+      if (val1[key] === undefined || val2[key] === undefined || !descriptor.type.equals(val1[key], val2[key])) {
+        return false;
       }
     }
     return true;
   }
 
   clone(val: T): T {
-    // TODO(demurgos): use Partial<T> once typedoc supports it
-    const result: any = {};
+    const result: Partial<T> = Object.create(null);
     for (const key in this.properties) {
-      const member: any = (<any> val)[key];
-      if (member !== undefined) {
-        result[key] = this.properties[key].type.clone(member);
-      }
+      result[key] = val[key] === undefined ? undefined : this.properties[key].type.clone(val[key]);
     }
     return result as T;
   }
 
-  diff(oldVal: T, newVal: T): Diff | undefined {
+  diff(oldVal: T, newVal: T): Diff<T> | undefined {
     let equal: boolean = true;
-    const result: Diff = {set: {}, unset: {}, update: {}};
+    const result: Diff<T> = {set: {}, unset: {}, update: {}};
     for (const key in this.properties) {
-      const descriptor: PropertyDescriptor<VersionedType<any, any, any, any>> = this.properties[key];
+      // TODO: Remove cast
+      const descriptor: PropertyDescriptor<VersionedType<any, any, any, any>> = <any> this.properties[key];
       const oldMember: any = (<any> oldVal)[key];
       const newMember: any = (<any> newVal)[key];
       if (oldMember !== undefined) {
@@ -253,32 +275,29 @@ export class DocumentType<T extends {}> implements VersionedType<T, json.Input, 
     return equal ? undefined : result;
   }
 
-  patch(oldVal: T, diff: Diff | undefined): T {
+  patch(oldVal: T, diff: Diff<T> | undefined): T {
     const result: T = this.clone(oldVal);
     if (diff === undefined) {
       return result;
     }
     for (const key in diff.set) {
-      (<any> result)[key] = this.properties[key].type.clone(diff.set[key]);
+      result[key] = this.properties[key].type.clone(diff.set[key]);
     }
     for (const key in diff.unset) {
-      delete (<any> result)[key];
+      Reflect.deleteProperty(result, key);
     }
     for (const key in diff.update) {
-      (<any> result)[key] = this.properties[key].type.patch((<any> result)[key][key], diff.update[key]);
+      // TODO: Remove cast
+      result[key] = (this.properties[key].type as any).patch(result[key], diff.update[key]);
     }
     return result;
   }
 
-  reverseDiff(diff: Diff | undefined): Diff | undefined {
+  reverseDiff(diff: Diff<T> | undefined): Diff<T> | undefined {
     if (diff === undefined) {
       return undefined;
     }
-    const result: Diff = {
-      set: {},
-      unset: {},
-      update: {},
-    };
+    const result: Diff<T> = {set: {}, unset: {}, update: {}};
     for (const key in diff.unset) {
       result.set[key] = this.properties[key].type.clone(diff.unset[key]);
     }
@@ -286,12 +305,13 @@ export class DocumentType<T extends {}> implements VersionedType<T, json.Input, 
       result.unset[key] = this.properties[key].type.clone(diff.set[key]);
     }
     for (const key in diff.update) {
-      result.update[key] = this.properties[key].type.reverseDiff(diff.update[key]);
+      // TODO: Remove cast
+      result.update[key] = (this.properties[key].type as any).reverseDiff(diff.update[key]);
     }
     return result;
   }
 
-  squash(diff1: Diff | undefined, diff2: Diff | undefined): Diff | undefined {
+  squash(diff1: Diff<T> | undefined, diff2: Diff<T> | undefined): Diff<T> | undefined {
     throw createNotImplementedError("DocumentType#squash");
   }
 
@@ -299,26 +319,31 @@ export class DocumentType<T extends {}> implements VersionedType<T, json.Input, 
     if (this._options === undefined) {
       throw new Incident("No pending options");
     }
-    const options: Options<VersionedType<any, any, any, any>> = typeof this._options === "function" ?
+    const options: DocumentTypeOptions<T> = typeof this._options === "function" ?
       this._options() :
       this._options;
 
     const ignoreExtraKeys: boolean = options.ignoreExtraKeys || false;
-    const properties: {[key: string]: PropertyDescriptor<VersionedType<any, any, any, any>>} = options.properties;
+    const properties: {[P in keyof T]: PropertyDescriptor<KryoType<any>>} = options.properties;
     const renameAll: CaseStyle | undefined = options.rename;
-    const keys: Map<string, string> = new Map<string, string>();
-    const outKeys: Map<string, string> = new Map<string, string>();
+    const keys: Map<keyof T, string> = renameKeys(properties, renameAll);
 
-    for (const key in properties) {
-      const renamed: string = renameAll === undefined ? key : rename(key, renameAll);
-      keys.set(key, renamed);
-      if (outKeys.has(renamed)) {
-        throw new Incident("NonBijectiveKeyRename", "Some keys are the same after renaming");
-      }
-      outKeys.set(renamed, key);
-    }
-
-    Object.assign(this, {ignoreExtraKeys, properties, rename: renameAll, keys, outKeys});
+    Object.assign(this, {ignoreExtraKeys, properties, rename: renameAll, keys});
     Object.freeze(this);
   }
+};
+
+export function renameKeys<T>(obj: T, renameAll?: CaseStyle): Map<keyof T, string> {
+  const keys: (keyof T)[] = Object.keys(obj) as (keyof T)[];
+  const result: Map<keyof T, string> = new Map();
+  const outKeys: Set<string> = new Set();
+  for (const key of keys) {
+    const renamed: string = renameAll === undefined ? key : rename(key, renameAll);
+    result.set(key, renamed);
+    if (outKeys.has(renamed)) {
+      throw new Incident("NonBijectiveKeyRename", "Some keys are the same after renaming");
+    }
+    outKeys.add(renamed);
+  }
+  return result;
 }