+++ src/lib/types/literal.ts
@@ -21,14 +21,25 @@ export interface Options<T, Output, Input extends Output, Diff> {
   value: T;
 }
 
+/**
+ * You may need to explicitly write the type or inference won't pick it.
+ * For example, in the case of enum values, inference will pick the type of the enum instead of
+ * the specific property you pass.
+ *
+ * @see https://github.com/Microsoft/TypeScript/issues/10195
+ */
 export class LiteralType<T> implements VersionedType<T, json.Input, json.Output, Diff> {
   readonly name: Name = name;
-  readonly type!: VersionedType<T, any, any, Diff>;
-  readonly value!: T;
+  readonly type: VersionedType<T, any, any, Diff>;
+  readonly value: T;
 
   private _options: Lazy<Options<T, any, any, any>>;
 
   constructor(options: Lazy<Options<T, any, any, any>>, lazy?: boolean) {
+    // TODO: Remove once TS 2.7 is better supported by editors
+    this.type = <any> undefined;
+    this.value = <any> undefined;
+
     this._options = options;
     if (lazy === undefined) {
       lazy = typeof options === "function";