+++ src/test/circular-types/fs-tree/file.ts
@@ -6,15 +6,15 @@ import { $FsNodeBase, FsNodeBase } from "./fs-node-base";
 import { $FsNodeType, FsNodeType } from "./fs-node-type";
 
 export interface File extends FsNodeBase {
-  type: FsNodeType.File;
+  tag: FsNodeType.File;
   size: number;
 }
 
-export const $File: DocumentType<File> = new DocumentType<File>({
+export const $File: DocumentType<File> = new DocumentType<File>(() => ({
   properties: {
     ...$FsNodeBase.properties,
-    type: {type: new LiteralType({type: $FsNodeType, value: FsNodeType.File})},
+    tag: {type: new LiteralType<FsNodeType.File>({type: $FsNodeType, value: FsNodeType.File})},
     size: {type: new IntegerType()},
   },
   rename: CaseStyle.SnakeCase,
-});
+}));