+++ src/test/circular-types/fs-tree/fs-node-base.ts
@@ -4,13 +4,13 @@ import { Ucs2StringType } from "../../../lib/types/ucs2-string";
 import { $FsNodeType, FsNodeType } from "./fs-node-type";
 
 export interface FsNodeBase {
-  type: FsNodeType;
+  tag: FsNodeType;
   name: string;
 }
 
 export const $FsNodeBase: DocumentType<FsNodeBase> = new DocumentType<FsNodeBase>({
   properties: {
-    type: {type: $FsNodeType},
+    tag: {type: $FsNodeType},
     name: {type: new Ucs2StringType({maxLength: Infinity})},
   },
   rename: CaseStyle.SnakeCase,